requirements.txt
slack-bolt==1.26.0
fastapi==0.121.1
uvicorn[standard]==0.38.0
httpx==0.28.1
python-dotenv==1.2.1
